requirements.txt
fastapi
uvicorn
python-multipart
requests
langchain-groq
openai
python-dotenv
streamlit
Pillow
httpx
numpy